Revert "Add DraggableList component"
