+++ package.json
@@ -109,7 +109,6 @@
     "prop-types": "^15.7.2",
     "pusher-js": "8.3.0",
     "react": "18.2.0",
-    "react-beautiful-dnd": "^13.1.1",
     "react-collapse": "^5.1.0",
     "react-content-loader": "^6.1.0",
     "react-dom": "18.1.0",
@@ -122,7 +121,6 @@
     "react-native-dev-menu": "^4.1.1",
     "react-native-device-info": "^10.3.0",
     "react-native-document-picker": "^8.0.0",
-    "react-native-draggable-flatlist": "^4.0.1",
     "react-native-fast-image": "^8.6.3",
     "react-native-fs": "^2.20.0",
     "react-native-gesture-handler": "2.12.0",
@@ -206,7 +204,6 @@
     "@types/mock-fs": "^4.13.1",
     "@types/pusher-js": "^5.1.0",
     "@types/react": "^18.2.12",
-    "@types/react-beautiful-dnd": "^13.1.4",
     "@types/react-collapse": "^5.0.1",
     "@types/react-dom": "^18.2.4",
     "@types/react-pdf": "^5.7.2",
