+++ src/libs/actions/Transaction.ts
@@ -31,8 +31,8 @@ function createInitialWaypoints(transactionID: string) {
     Onyx.merge(`${ONYXKEYS.COLLECTION.TRANSACTION}${transactionID}`, {
         comment: {
             waypoints: {
-                waypoint0: null,
-                waypoint1: null,
+                waypoint0: {},
+                waypoint1: {},
             },
         },
     });
@@ -201,34 +201,4 @@ function getRoute(transactionID: string, waypoints: WaypointCollection) {
     );
 }
 
-/**
- * Updates all waypoints stored in the transaction specified by the provided transactionID.
- *
- * @param transactionID - The ID of the transaction to be updated
- * @param waypoints - An object containing all the waypoints
- *                             which will replace the existing ones.
- */
-function updateWaypoints(transactionID: string, waypoints: WaypointCollection): Promise<void> {
-    return Onyx.merge(`${ONYXKEYS.COLLECTION.TRANSACTION}${transactionID}`, {
-        comment: {
-            waypoints,
-        },
-
-        // Empty out errors when we're saving new waypoints as this indicates the user is updating their input
-        errorFields: {
-            route: null,
-        },
-
-        // Clear the existing route so that we don't show an old route
-        routes: {
-            route0: {
-                distance: null,
-                geometry: {
-                    coordinates: null,
-                },
-            },
-        },
-    });
-}
-
-export {addStop, createInitialWaypoints, saveWaypoint, removeWaypoint, getRoute, updateWaypoints};
+export {addStop, createInitialWaypoints, saveWaypoint, removeWaypoint, getRoute};
