+++ src/styles/styles.js
@@ -3631,6 +3631,12 @@ const styles = (theme) => ({
         maxWidth: 400,
     },
 
+    distanceRequestContainer: (maxHeight) => ({
+        ...flex.flexShrink2,
+        minHeight: variables.optionRowHeight * 2,
+        maxHeight,
+    }),
+
     mapViewContainer: {
         ...flex.flex1,
         ...spacing.p4,
